requirements.txt
appdirs==1.4.3
asgiref==3.2.10
beautifulsoup4==4.8.2
CacheControl==0.12.6
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.1.1
django-environ==0.4.5
django-filter==2.3.0
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
gunicorn==20.0.4
html5lib==1.1
idna==2.10
ipaddr==2.2.0
jdcal==1.4.1
l18n==2018.5
lockfile==0.12.2
msgpack==0.6.2
openpyxl==3.0.5
packaging==20.3
pep517==0.8.2
Pillow==7.2.0
progress==1.5
psycopg2-binary==2.8.6
pyparsing==2.4.6
pytoml==0.1.21
pytz==2020.1
requests==2.24.0
retrying==1.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.25.10
wagtail==2.10.1
webencodings==0.5.1
Willow==1.4
xlrd==1.2.0
XlsxWriter==1.3.3
xlwt==1.3.0
